fix: changed location of files
