--- src/Consts/exercisesRS.js
@@ -1,49 +1,0 @@
-const klasy = {
-  class0: {
-    0.1: 'Socjalizacja',
-    0.2: 'Waruj/Siad przez 1 minutę',
-    0.3: 'Chodzenie przy nodze',
-    0.4: 'Próba trzymania aportu',
-    0.5: 'Przywołanie',
-    0.6: 'Obieganie 3 pachołków z przywołaniem',
-    0.7: 'Zmiana pozycji',
-    0.8: 'Wrażenie ogólne',
-  },
-
-  class1: {
-    1.1: 'Siad w grupie przez 1 minutę, przewodnik w zasięgu wzroku',
-    1.2: 'Chodzenie przy nodze',
-    1.3: 'Stój lub siad lub waruj w marszu',
-    1.4: 'Przywołanie',
-    1.5: 'Wysyłanie psa do kwadratu',
-    1.6: 'Zmiany pozycji na odległość',
-    1.7: 'Aport koziołka i skok przez przeszkodę',
-    1.8: 'Obieganie grupy pachołkóww / walca z przywołaniem',
-    1.9: 'Wrażenie ogólne',
-  },
-
-  class2: {
-    2.01: 'Waruj w grupie przez 2 minuty, przewodnik poza zasięgiem wzroku',
-    2.2: 'Chodzenie przy nodze',
-    2.3: 'Stój/siad/waruj w marszu',
-    2.4: 'Przywołanie z zatrzymaniem stój',
-    2.5: 'Wysyłanie psa kierunkowe, do kwadratu, waruj i przywołanie',
-    2.6: 'Aport kierunkowy',
-    2.7: 'Rozpoznanie i przyniesienie własnego przedmiotu',
-    2.8: 'Zmiany pozycji na odległość',
-    2.9: 'Wysyłanie wokół grupy pachołków/walca, zatrzymanie i skok przez przeszkodę',
-    2.1: 'Wrażenie ogólne',
-  },
-  class3: {
-    3.01: 'Siad w grupie przez 2 minuty, przewodnik poza zasięgiem wzroku',
-    3.2: 'Waruj w grupie przez 1 minutę i przywołanie',
-    3.3: 'Chodzenie przy nodze',
-    3.4: 'Stój, siad i waruj w marszu',
-    3.5: 'Przywołanie z zatrzymaniem stój/siad/waruj',
-    3.6: 'Wysyłanie psa kierunkowe, do kwadratu, waruj i przywołanie',
-    3.7: 'Aport kierunkowy',
-    3.8: 'Wysyłanie wokół grupy pachołków, zatrzymanie (stój/siad/waruj), aport koziołka i skok przez przeszkodę',
-    3.9: 'Rozpoznanie i przyniesienie własnego przedmiotu',
-    3.1: 'Zmiany pozycji na odległość',
-  },
-};
