--- src/Data/Dummy-data/leaderboardClassResults.js
@@ -1,43 +1,0 @@
-// dummy data representing results in one of selected classes
-// array of objects
-
-const LeaderboardClassResult = [
-  [
-    // class0 results
-    { text: 'Woof', score: 310 },
-    { text: 'Pjes', score: 300 },
-    { text: 'Ugryź', score: 170 },
-    { text: 'Baseji', score: 220 },
-    { text: 'Tomato', score: 120 },
-    { text: 'Wariat', score: 150 },
-  ],
-  [
-    // class1 results
-    { text: 'Woof', score: 310 },
-    { text: 'Pjes', score: 300 },
-    { text: 'Ugryź', score: 170 },
-    { text: 'Baseji', score: 220 },
-    { text: 'Tomato', score: 120 },
-    { text: 'Wariat', score: 150 },
-  ],
-  [
-    // class2 results
-    { text: 'Woof', score: 310 },
-    { text: 'Pjes', score: 300 },
-    { text: 'Ugryź', score: 170 },
-    { text: 'Baseji', score: 220 },
-    { text: 'Tomato', score: 120 },
-    { text: 'Wariat', score: 150 },
-  ],
-  [
-    // class3 results
-    { text: 'Woof', score: 310 },
-    { text: 'Pjes', score: 300 },
-    { text: 'Ugryź', score: 170 },
-    { text: 'Baseji', score: 220 },
-    { text: 'Tomato', score: 120 },
-    { text: 'Wariat', score: 150 },
-  ],
-];
-
-export default LeaderboardClassResult;
