+++ src/PagesBody/DogSummaryPage/DogSummaryPage.js
@@ -7,15 +7,14 @@ import WYNIK_RANDOMOWEGO_PSA from './../../Data/Dummy-data/RandomDogResult';
 
 const sendDataHandler = () => {};
 
-// const fakeId = 'a0347677-c3c9-4edc-9d46-fed4a958fdc2';
-// const fakeClass = '1';
-// const fakeDogName = 'Kluska';
-
 const DogSummaryPage = () => {
   return (
     <ColumnWrapper>
-      <LeaderboardList result={WYNIK_RANDOMOWEGO_PSA} />
-      {/* <LeaderboardList contestId={fakeId} classNumber={fakeClass} dogName={fakeDogName}/> */}
+      <LeaderboardList
+        contestId={'a0347677-c3c9-4edc-9d46-fed4a958fdc2'}
+        classId={'0'}
+        dogName={'Woof'}
+      />
       <PenaltyOrDisqualifiedLine result={WYNIK_RANDOMOWEGO_PSA} />
       <SummaryLine result={WYNIK_RANDOMOWEGO_PSA} />
       <ColumnWrapper paddingLeftRight={1}>
